권한 처리

+++ src/main/java/shop/miraclecoding/blog/board/BoardController.java
@@ -8,9 +8,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.miraclecoding.blog._core.errors.exception.Exception400;
+import shop.miraclecoding.blog._core.errors.exception.Exception401;
+import shop.miraclecoding.blog._core.errors.exception.Exception403;
+import shop.miraclecoding.blog._core.errors.exception.Exception404;
 import shop.miraclecoding.blog.user.User;
 import shop.miraclecoding.blog.user.UserRequest;
 
+import java.io.Writer;
 import java.util.List;
 
 @Controller
@@ -35,6 +40,11 @@ public class BoardController {
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest request){
         Board board = boardRepository.findById(id);
+
+        if (board == null){
+            throw new Exception404("해당 게시글을 찾을 수 없습니다.");
+        }
+
         request.setAttribute("board",board);
         return "board/update-form";
     }
@@ -42,20 +52,44 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정할 권한이 없습니다!");
+        }
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/"+id;
     }
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 삭제할 권한이 없습니다!");
+        }
+
         boardRepository.deleteById(id);
         return "redirect:/";
     }
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
-        request.setAttribute("board",board);
+
+        // 로그인을 했고, 게시글의 작성자면 isAuthor = true
+        boolean isAuthor = false;
+        if (sessionUser != null) {
+            if (sessionUser.getId() == board.getUser().getId()) {
+                isAuthor = true;
+            }
+        }
+
+        request.setAttribute("isAuthor", isAuthor);
+        request.setAttribute("board", board);
         return "board/detail";
     }
